requirements.txt
Django==2.1.5

pinax-likes
django-crispy-forms
psycopg2
django-cors-headers
stream_django
django-autoslug
pillow
gunicorn
django-allauth
django-initial-avatars
djangorestframework
pinax-messages
sentry-sdk
pyuploadcare
django-tinymce4-lite
sorl-thumbnail
django-el-pagination
django-random-queryset
django-querysetsequence